JUnitExamples/JUnitExamples/src/main/java/com/example/junitexamples/ArrayUtils.java
package com.example.junitexamples;

public class ArrayUtils {
    public static int findSmallestNumber(int[] numbers) {
        if (numbers == null || numbers.length == 0) {
            throw new IllegalArgumentException("De array mag niet leeg zijn.");
        }

        int smallest = numbers[0];
        for (int i = 1; i < numbers.length; i++) {
            if (numbers[i] < smallest) {
                smallest = numbers[i];
            }
        }

        return smallest;
    }
}
